New option to escape DNS names
Option --rfc1035 escapes unusual characters in DNS names.
Everything outside of [a-zA-Z0-9_-] is escaped like \012.

A shorter notation for special characters like \. \\ \" etc.
is not used on purpose because packetq stores decoded names as strings.
Usage of \. would break parsing names with embedded dot using
the internal rsplit() function.
Also, users would have to know which special characters are escaped and
which are not and that would be very very error prone e.g. when writing
WHERE expressions.

The original output mode replaces non-printable characters with
$ character is still used by default for compatibility.
This original mode does not allow a correct analysis of names with
embedded dots, but that is "an old feature."